Break up babel monitor
This patch breaks up babel monitor into three distinct files, one for
the structs, one for the syncronous parsing code and one for the actual
request code that interacts with the tcp streams.
