+++ rita_client/src/error.rs
@@ -8,7 +8,7 @@ use std::{
 
 use althea_kernel_interface::KernelInterfaceError;
 use awc::error::{JsonPayloadError, SendRequestError};
-use babel_monitor::BabelMonitorError;
+use babel_monitor::structs::BabelMonitorError;
 use compressed_log::builder::LoggerError;
 use log::SetLoggerError;
 use rita_common::RitaCommonError;
