+++ rita_common/src/tunnel_manager/gc.rs
@@ -3,9 +3,9 @@ use crate::tunnel_manager::TUNNEL_MANAGER;
 use crate::RitaCommonError;
 use crate::KI;
 use althea_types::Identity;
-use babel_monitor::Interface;
+use babel_monitor::structs::Interface;
+use std::time::Duration;
 use std::{collections::HashMap, time::Instant};
-use std::{time::Duration};
 
 /// Performs a cleanup of all babel tunnels that we have not heard from in the configured time
 /// tunnel_timeout:
@@ -24,7 +24,10 @@ use std::{time::Duration};
 /// 'up' we will gc it for recreation via the normal hello/ihu process, this prevents us
 /// from having tunnels that don't work for babel peers
 pub fn tm_trigger_gc(
-    tunnel_timeout: Duration, tunnel_handshake_timeout: Duration, babel_interfaces: Vec<Interface>) -> Result<(), RitaCommonError> {
+    tunnel_timeout: Duration,
+    tunnel_handshake_timeout: Duration,
+    babel_interfaces: Vec<Interface>,
+) -> Result<(), RitaCommonError> {
     let tunnel_manager = &mut *TUNNEL_MANAGER.write().unwrap();
 
     let interfaces = into_interfaces_hashmap(&babel_interfaces);
@@ -36,7 +39,12 @@ pub fn tm_trigger_gc(
     // checker issues, we should consider a method that does modify in place
     for (_identity, tunnels) in tunnel_manager.tunnels.iter() {
         for tunnel in tunnels.iter() {
-            if tunnel_should_be_kept(tunnel, tunnel_handshake_timeout, tunnel_timeout, &interfaces) {
+            if tunnel_should_be_kept(
+                tunnel,
+                tunnel_handshake_timeout,
+                tunnel_timeout,
+                &interfaces,
+            ) {
                 insert_into_tunnel_list(tunnel, &mut good);
             } else {
                 insert_into_tunnel_list(tunnel, &mut to_delete)
@@ -131,8 +139,7 @@ fn tunnel_should_be_kept(
     // this is almost always true, unless one of the two is in the future versus 'now' it's safe to just skip this
     // for the next gc round in that case.
     if let (Some(since_created), Some(since_last_contact)) = (since_created, since_last_contact) {
-        let handshake_timeout =
-            !check_handshake_time(tunnel_handshake_timeout, &tunnel.iface_name);
+        let handshake_timeout = !check_handshake_time(tunnel_handshake_timeout, &tunnel.iface_name);
         let created_recently = since_created < tunnel_timeout;
         let tunnel_up = tunnel_up(interfaces, &tunnel.iface_name);
         let contact_timeout = since_last_contact > tunnel_timeout;
