+++ rita_common/src/tunnel_manager/mod.rs
@@ -20,9 +20,9 @@ use althea_types::Identity;
 use althea_types::LocalIdentity;
 use babel_monitor::monitor;
 use babel_monitor::open_babel_stream;
+use babel_monitor::structs::BabelMonitorError;
+use babel_monitor::structs::Interface;
 use babel_monitor::unmonitor;
-use babel_monitor::BabelMonitorError;
-use babel_monitor::Interface;
 use std::collections::HashMap;
 use std::collections::HashSet;
 use std::collections::VecDeque;
